% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{add_node}
\alias{add_node}
\title{Add a node to a nodeset}
\usage{
add_node(nodeset, id)
}
\arguments{
\item{nodeset}{A `threadle_nodeset` object.}

\item{id}{Node ID.}
}
\value{
CLI output.
}
\description{
Add a node to a nodeset
}
