% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{create_network}
\alias{create_network}
\title{Create a new network in Threadle}
\usage{
create_network(name, nodeset, label = NULL)
}
\arguments{
\item{name}{Name of the network variable.}

\item{nodeset}{A `threadle_nodeset` object.}

\item{label}{Optional display label.}
}
\value{
A `threadle_network` object.
}
\description{
Create a new network in Threadle
}
