% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{create_nodeset}
\alias{create_nodeset}
\title{Create a new nodeset in Threadle}
\usage{
create_nodeset(name)
}
\arguments{
\item{name}{Name of the R variable to assign in the CLI environment.}
}
\value{
A `threadle_nodeset` object.
}
\description{
Create a new nodeset in Threadle
}
