% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{define_attr}
\alias{define_attr}
\title{Define an attribute for a nodeset}
\usage{
define_attr(nodeset, attrname, attrtype)
}
\arguments{
\item{nodeset}{A `threadle_nodeset` object.}

\item{attrname}{Name of the attribute.}

\item{attrtype}{Attribute type.}
}
\value{
CLI output.
}
\description{
Define an attribute for a nodeset
}
