% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{define_layer}
\alias{define_layer}
\title{Define a layer in a network}
\usage{
define_layer(
  network,
  layername,
  mode,
  directed = FALSE,
  valuetype = "binary",
  selfties = FALSE
)
}
\arguments{
\item{network}{A `threadle_network` object.}

\item{layername}{Name of the layer.}

\item{mode}{Layer mode (1 or 2).}

\item{directed}{Logical; whether ties are directed.}

\item{valuetype}{"binary" or "valued".}

\item{selfties}{Logical; whether self-ties are allowed.}
}
\value{
CLI output.
}
\description{
Define a layer in a network
}
