% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{get_attr}
\alias{get_attr}
\title{Get the value of a node attribute}
\usage{
get_attr(nodeset, nodeid, attrname)
}
\arguments{
\item{nodeset}{Nodeset name or object.}

\item{nodeid}{Node ID.}

\item{attrname}{Attribute name.}
}
\value{
CLI output.
}
\description{
Get the value of a node attribute
}
