% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{get_nbr_nodes}
\alias{get_nbr_nodes}
\title{Get the number of nodes in a structure}
\usage{
get_nbr_nodes(name)
}
\arguments{
\item{name}{Name of the structure.}
}
\value{
A numeric value.
}
\description{
Get the number of nodes in a structure
}
