% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{get_node_alters}
\alias{get_node_alters}
\title{Get alters of a node within a network layer}
\usage{
get_node_alters(name, layername, nodeid, direction = "both")
}
\arguments{
\item{name}{Name of the network.}

\item{layername}{Layer to query.}

\item{nodeid}{Node ID.}

\item{direction}{Tie direction ("both", "in", "out").}
}
\value{
Parsed JSON list of alters.
}
\description{
Get alters of a node within a network layer
}
