% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{get_nodeid_by_index}
\alias{get_nodeid_by_index}
\title{Get a node ID by index}
\usage{
get_nodeid_by_index(name, index)
}
\arguments{
\item{name}{Name of the structure.}

\item{index}{Numeric index.}
}
\value{
The node ID.
}
\description{
Get a node ID by index
}
