% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{get_random_alter}
\alias{get_random_alter}
\title{Get a random alter for a node}
\usage{
get_random_alter(
  name,
  nodeid,
  layername = "",
  direction = "both",
  balanced = "false"
)
}
\arguments{
\item{name}{Network name.}

\item{nodeid}{Node ID.}

\item{layername}{Optional layer.}

\item{direction}{Direction ("both", "in", "out").}

\item{balanced}{Whether selection should be balanced.}
}
\value{
A node ID (numeric).
}
\description{
Get a random alter for a node
}
