% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{get_random_node}
\alias{get_random_node}
\title{Get a random node from a structure}
\usage{
get_random_node(name)
}
\arguments{
\item{name}{Structure name.}
}
\value{
A node ID (numeric).
}
\description{
Get a random node from a structure
}
