% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{get_workdir}
\alias{get_workdir}
\title{Get the current working directory from Threadle CLI}
\usage{
get_workdir()
}
\value{
The working directory as returned by Threadle.
}
\description{
Get the current working directory from Threadle CLI
}
