% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{info}
\alias{info}
\title{Retrieve information from a Threadle object}
\usage{
info(name, format = "json")
}
\arguments{
\item{name}{Name of the structure.}

\item{format}{Output format ("json" or raw text).}
}
\value{
Parsed JSON (list) or raw CLI text.
}
\description{
Retrieve information from a Threadle object
}
