% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{inventory}
\alias{inventory}
\title{List all objects currently defined in Threadle}
\usage{
inventory(format = "json")
}
\arguments{
\item{format}{Output format ("json" or raw text).}
}
\value{
Parsed JSON or raw text.
}
\description{
List all objects currently defined in Threadle
}
