% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{load_file}
\alias{load_file}
\title{Load a file into Threadle}
\usage{
load_file(name, file, type)
}
\arguments{
\item{name}{Name for the resulting object.}

\item{file}{File path.}

\item{type}{Type of structure (e.g. "network", "nodeset").}
}
\value{
An object with class corresponding to the loaded type.
}
\description{
Load a file into Threadle
}
