% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{load_network}
\alias{load_network}
\title{Load a network structure from a file}
\usage{
load_network(name, file)
}
\arguments{
\item{name}{Name for the network object.}

\item{file}{Path to the network file.}
}
\value{
A `threadle_network` object.
}
\description{
Load a network structure from a file
}
