% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{op-null-coalesce}
\alias{op-null-coalesce}
\alias{\%||\%}
\title{Null-coalescing operator}
\usage{
a \%||\% b
}
\arguments{
\item{a}{First value.}

\item{b}{Fallback value.}
}
\value{
\code{a} if non-\code{NULL}, else \code{b}.
}
\description{
Returns \code{a} if it is not \code{NULL}, otherwise returns \code{b}.
}
\keyword{internal}
