% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{set_attr}
\alias{set_attr}
\title{Set the value of a node attribute}
\usage{
set_attr(nodeset, nodeid, attrname, attrvalue)
}
\arguments{
\item{nodeset}{A `threadle_nodeset` object.}

\item{nodeid}{Node ID.}

\item{attrname}{Attribute name.}

\item{attrvalue}{Value to assign.}
}
\value{
CLI output.
}
\description{
Set the value of a node attribute
}
