% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_add_aff}
\alias{th_add_aff}
\title{Add an affiliation to a two-mode layer}
\usage{
th_add_aff(
  network,
  layername,
  nodeId,
  hypername,
  addmissingnode = TRUE,
  addmissingaffiliation = TRUE
)
}
\arguments{
\item{network}{A \code{threadle_network} object or a character string giving
the name of a network in the Threadle CLI environment.}

\item{layername}{Layer name.}

\item{nodeId}{Node ID.}

\item{hypername}{Name of the hyperedge.}

\item{addmissingnode}{Logical; if \code{TRUE}, missing nodes are created automatically.}

\item{addmissingaffiliation}{Logical; if \code{TRUE}, missing affiliations are created automatically.}
}
\value{
\code{NULL}, invisibly.
}
\description{
\code{th_add_aff()} adds an affiliation (hyperedge membership) between a node and a hyperedge in a two-mode layer.
}
