% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_add_hyper}
\alias{th_add_hyper}
\title{Add a hyperedge to a two-mode layer}
\usage{
th_add_hyper(
  network,
  layername,
  hypername,
  nodes = c(),
  addmissingnodes = TRUE
)
}
\arguments{
\item{network}{A \code{threadle_network} object or a character string giving
the name of a network in the Threadle CLI environment.}

\item{layername}{Layer name.}

\item{hypername}{Name of the hyperedge.}

\item{nodes}{Vector of node to attach to the hyperedge. If \code{NULL},
creates the hyperedge without adding nodes.}

\item{addmissingnodes}{Logical; if \code{TRUE}, missing nodes are created and added. Defaults to \code{TRUE}.}
}
\value{
\code{NULL}, invisibly.
}
\description{
\code{th_add_hyper()} adds a hyperedge (affiliation) to the hyperedge set of a 2-mode layer.
}
\details{
If a hyperedge with the same name already exists, it is replaced. Duplicate node IDs in \code{nodes} are ignored.
}
\examples{
\dontshow{if (th_is_available()) withAutoprint(\{ # examplesIf}
th_start_threadle()

ns <- th_create_nodeset("ns", createnodes = 5)
net <- th_create_network("net", ns)
th_add_layer(net, "t2", mode = 2)
th_add_hyper(net, "t2", "group1", nodes = c(1, 2))
th_get_node_hyperedges(net, "t2", nodeid = 1)
th_stop_threadle()
\dontshow{\}) # examplesIf}
}
