% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_add_layer}
\alias{th_add_layer}
\title{Add a relational layer in a network}
\usage{
th_add_layer(
  network,
  layername,
  mode,
  directed = FALSE,
  valuetype = c("binary", "valued"),
  selfties = FALSE
)
}
\arguments{
\item{network}{A \code{threadle_network} object or a character string giving
the name of a network in the Threadle CLI environment.}

\item{layername}{Layer name.}

\item{mode}{Layer mode (1 or 2).}

\item{directed}{Logical; whether ties are directed (only for 1-mode layers).}

\item{valuetype}{"binary" or "valued" (only for 1-mode layers).}

\item{selfties}{Logical; whether self-ties are allowed (only for 1-mode layers).}
}
\value{
\code{NULL}, invisibly.
}
\description{
\code{th_add_layer()} adds a new layer to a network, which can be either one-mode or two-mode.
}
\examples{
\dontshow{if (th_is_available()) withAutoprint(\{ # examplesIf}
th_start_threadle()

ns <- th_create_nodeset("ns", createnodes = 5)
net <- th_create_network("net", ns)
th_add_layer(net, "l1", mode = 1)
th_add_layer(net, "t2", mode = 2)
th_info(net)
th_stop_threadle()
\dontshow{\}) # examplesIf}
}
