% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_clear_layer}
\alias{th_clear_layer}
\title{Remove all edges in a layer}
\usage{
th_clear_layer(network, layername)
}
\arguments{
\item{network}{A \code{threadle_network} object or a character string giving
the name of a network in the Threadle CLI environment.}

\item{layername}{Layer name.}
}
\value{
\code{NULL}, invisibly.
}
\description{
\code{th_clear_layer()} removes all edges in a layer but keep the layer.
For two-mode layers, all hyperedges are removed.
}
