% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_cli_help}
\alias{th_cli_help}
\title{Help from Threadle backend (CLI), not R package documentation.}
\usage{
th_cli_help(arg0 = NULL, file = NULL)
}
\description{
Help from Threadle backend (CLI), not R package documentation.
}
