% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_create_network}
\alias{th_create_network}
\title{Create a network using the provided nodeset}
\usage{
th_create_network(var, nodeset, name = NULL)
}
\arguments{
\item{var}{Name of the assigned variable in the Threadle CLI environment.}

\item{nodeset}{A \code{threadle_nodeset} object or a character string giving
the name of a nodeset in the Threadle CLI environment.}

\item{name}{Optional internal name stored in the Threadle object metadata.
This does not affect how the object is referenced in commands; commands
use 'var'.}
}
\value{
A \code{threadle_network} object.
}
\description{
\code{th_create_network()} creates a new network from an existing nodeset.
}
\examples{
\dontshow{if (th_is_available()) withAutoprint(\{ # examplesIf}
th_start_threadle()

ns <- th_create_nodeset("ns", createnodes = 5)
net <- th_create_network("net", ns)
th_info(net)
th_stop_threadle()
\dontshow{\}) # examplesIf}
}
