% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_define_attr}
\alias{th_define_attr}
\title{Define an attribute for a nodeset}
\usage{
th_define_attr(
  structure,
  attrname,
  attrtype = c("int", "char", "float", "bool")
)
}
\arguments{
\item{structure}{A \code{threadle_nodeset} or \code{threadle_network} object, or a character
string naming a structure in the Threadle CLI environment.}

\item{attrname}{Attribute name.}

\item{attrtype}{Attribute type ("int","char", "float", or "bool")}
}
\value{
\code{NULL}, invisibly.
}
\description{
\code{th_define_attr()} defines a node attribute for a nodeset or the nodeset of the provided network.
}
\examples{
\dontshow{if (th_is_available()) withAutoprint(\{ # examplesIf}
th_start_threadle()

ns <- th_create_nodeset("ns", createnodes = 3)
th_define_attr(ns, "score", "int")
th_set_attr(ns, nodeid = 1, attrname = "score", attrvalue = 10)
th_get_attr(ns, nodeid = 1, attrname = "score")
th_stop_threadle()
\dontshow{\}) # examplesIf}
}
