% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_density}
\alias{th_density}
\title{Calculate the density of a layer}
\usage{
th_density(network, layername)
}
\arguments{
\item{network}{A \code{threadle_network} object or a character string giving
the name of a network in the Threadle CLI environment.}

\item{layername}{Name of the layer for which density is computed.}
}
\value{
A numeric scalar giving the layer density.
}
\description{
\code{th_density()} computes the density of a layer, treating all existing ties as binary.
}
