% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_filter}
\alias{th_filter}
\title{Create a nodeset by a node attribute}
\usage{
th_filter(name, nodeset, attrname, cond, attrvalue)
}
\arguments{
\item{name}{Name of the new nodeset variable}

\item{nodeset}{A \code{threadle_nodeset} object or a character string giving
the name of a nodeset in the Threadle CLI environment.}

\item{attrname}{Name of the node attribute to filter on.}

\item{cond}{Condition operator: \code{"eq"}, \code{"ne"}, \code{"gt"}, \code{"lt"}, \code{"ge"}, \code{"le"},
\code{"isnull"}, or \code{"notnull"}.}

\item{attrvalue}{Attribute value used for the condition.}
}
\value{
A \code{threadle_nodeset} object.
}
\description{
\code{th_filter()} creates a new nodeset containing the nodes that satisfy a condition on a specified
node attribute.
}
\examples{
\dontshow{if (th_is_available()) withAutoprint(\{ # examplesIf}
th_start_threadle()

ns <- th_create_nodeset("ns", createnodes = 4)
th_define_attr(ns, "score", "int")
th_set_attr(ns, nodeid = 0, "score", 10)
th_set_attr(ns, nodeid = 1, "score", 20)
th_set_attr(ns, nodeid = 2, "score", 30)
th_set_attr(ns, nodeid = 3, "score", 40)
hi <- th_filter("hi", ns, attrname = "score", cond = "ge", attrvalue = 30)
th_get_all_nodes(hi, offset = 0, limit = 10)
th_stop_threadle()
\dontshow{\}) # examplesIf}
}
