% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_get_all_nodes}
\alias{th_get_all_nodes}
\title{Get nodes from a nodeset}
\usage{
th_get_all_nodes(structure, offset, limit = 1000)
}
\arguments{
\item{structure}{A \code{threadle_nodeset} or \code{threadle_network} object, or a character string
naming a structure in the Threadle CLI environment.}

\item{offset}{Starting index (0-based). Defaults to \code{0}.}

\item{limit}{Maximum number of nodes to return. Defaults to 1000.}
}
\value{
An integer vector of node IDs.
}
\description{
\code{th_get_all_nodes} gets node IDs from a nodeset (or the nodeset of the provided network).
}
