% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_get_node_alters}
\alias{th_get_node_alters}
\title{Get alters of a node within a network layer}
\usage{
th_get_node_alters(
  network,
  nodeid,
  layername = "",
  direction = c("both", "in", "out"),
  unique = FALSE
)
}
\arguments{
\item{network}{Name of the network.}

\item{nodeid}{Node ID.}

\item{layername}{Layer to query.}

\item{direction}{Tie direction ("both", "in", "out").}
}
\value{
Parsed JSON list of alters.
}
\description{
Get alters of a node within a network layer
}
