% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_get_node_hyperedges}
\alias{th_get_node_hyperedges}
\title{Get hyperedges affiliated with a node}
\usage{
th_get_node_hyperedges(network, layername, nodeid)
}
\arguments{
\item{network}{A \code{threadle_network} object or a character string giving
the name of a network in the Threadle CLI environment.}

\item{layername}{Name of the two-mode layer to query.}

\item{nodeid}{Node ID.}
}
\value{
A character vector of hyperedge names.
}
\description{
\code{th_get_node_hyperedges()} gets the names of hyperedges
that a node is affiliated with in a two-mode layer.
}
