% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_get_nodeid_by_index}
\alias{th_get_nodeid_by_index}
\title{Get node ID by index position}
\usage{
th_get_nodeid_by_index(structure, index)
}
\arguments{
\item{structure}{A \code{threadle_nodeset} or \code{threadle_network} object, or a character
string naming a structure in the Threadle CLI environment.}

\item{index}{Zero-based index position of the node in the nodeset.}
}
\value{
An integer scalar giving the node ID at the requested index position.
}
\description{
\code{th_get_nodeid_by_index()} gets the node ID at a specified index position
in a nodeset (or the nodeset of the provided network).
}
\examples{
\dontshow{if (th_is_available()) withAutoprint(\{ # examplesIf}
th_start_threadle()

ns <- th_create_nodeset("ns", createnodes = 3)
th_get_nodeid_by_index(ns, index = 0)
th_stop_threadle()
\dontshow{\}) # examplesIf}
}
