% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_get_random_alter}
\alias{th_get_random_alter}
\title{Get a random alter of a node}
\usage{
th_get_random_alter(
  network,
  nodeid,
  layername = "",
  direction = "both",
  balanced = FALSE
)
}
\arguments{
\item{network}{A \code{threadle_network} object or a character string giving the name of
a network in the Threadle CLI environment.}

\item{nodeid}{Node ID.}

\item{layername}{Optional layer name. If \code{NULL}, alters are considered across all layers.}

\item{direction}{Which ties to count: \code{"both"} (default), \code{"in"}, or \code{"out"}.}

\item{balanced}{Logical; only used when \code{layername} is \code{NULL}.}
}
\value{
An integer scalar giving the node ID of the sampled alter.
}
\description{
\code{th_get_random_alter()} gets the node ID of a randomly selected alter of a given node.
}
\examples{
\dontshow{if (th_is_available()) withAutoprint(\{ # examplesIf}
th_start_threadle()

ns <- th_create_nodeset("ns", createnodes = 4)
net <- th_create_network("net", ns)
th_add_layer(net, "l1", mode = 1)
th_add_edge(net, "l1", node1id = 1, node2id = 2)
th_add_edge(net, "l1", node1id = 1, node2id = 3)
th_get_random_alter(net, nodeid = 1, layername = "l1", direction = "both")
th_stop_threadle()
\dontshow{\}) # examplesIf}
}
