% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_get_random_node}
\alias{th_get_random_node}
\title{Get a random node ID}
\usage{
th_get_random_node(structure)
}
\arguments{
\item{structure}{A \code{threadle_nodeset} or \code{threadle_network} object, or a character
string naming a structure in the Threadle CLI environment.}
}
\value{
An integer scalar giving the sampled node ID.
}
\description{
\code{th_get_random_node()} gets a randomly selected node ID from a nodeset (or the nodeset of the provided network).
}
