% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_get_workdir}
\alias{th_get_workdir}
\title{Get Threadle working directory}
\usage{
th_get_workdir()
}
\value{
A character scalar giving the current working directory used by Threadle.
}
\description{
\code{th_get_workdir} gets the working directory currently used by the Threadle.
}
\examples{
\dontshow{if (th_is_available()) withAutoprint(\{ # examplesIf}
th_start_threadle()

th_get_workdir()
th_stop_threadle()
\dontshow{\}) # examplesIf}
}
