% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_i}
\alias{th_i}
\title{List all stored data objects}
\usage{
th_i()
}
\value{
A named list mapping object names to their types.
}
\description{
\code{th_i()} lists an inventory of the data objects currently stored in the Threadle.
}
\details{
This command is in honor of all 1970's text adventure games,
where 'i' was used to check what you were carrying.
}
\examples{
\dontshow{if (th_is_available()) withAutoprint(\{ # examplesIf}
th_start_threadle()

th_i()
th_stop_threadle()
\dontshow{\}) # examplesIf}
}
