% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_info}
\alias{th_info}
\title{Get structure metadata}
\usage{
th_info(structure)
}
\arguments{
\item{structure}{A \code{threadle_nodeset} or \code{threadle_network} object, or a character
string naming a structure in the Threadle CLI environment.}
}
\value{
A named list containing structure metadata.
}
\description{
\code{th_info()} gets metadata about a nodeset or network.
}
\examples{
\dontshow{if (th_is_available()) withAutoprint(\{ # examplesIf}
th_start_threadle()

ns <- th_create_nodeset("ns", createnodes = 3)
th_info(ns)
th_stop_threadle()
\dontshow{\}) # examplesIf}
}
