% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_inventory}
\alias{th_inventory}
\title{List all objects currently stored as variables in Threadle}
\usage{
th_inventory()
}
\value{
Parsed JSON or raw text.
}
\description{
List all objects currently stored as variables in Threadle
}
