% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_load_example}
\alias{th_load_example}
\title{Load the bundled example network and nodeset into Threadle}
\usage{
th_load_example(name = "mynet", folder = "threadle_examples")
}
\description{
Load the bundled example network and nodeset into Threadle
}
