% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_load_script}
\alias{th_load_script}
\title{Load and run a Threadle script}
\usage{
th_load_script(file)
}
\arguments{
\item{file}{Path to a script file containing CLI commands.}
}
\value{
\code{NULL}, invisibly.
}
\description{
\code{th_load_script()} loads and executes a script file containing Threadle CLI commands.
}
\details{
Execution aborts if an error is encountered.
}
