% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_preview}
\alias{th_preview}
\title{Preview a structure}
\usage{
th_preview(structure, maxlines = 50)
}
\arguments{
\item{structure}{A \code{threadle_nodeset} or \code{threadle_network} object, or a character
string naming a structure in the Threadle CLI environment.}

\item{maxlines}{Maximum number of lines to output. Defaults to \code{50}.}
}
\value{
A character vector of preview lines.
}
\description{
\code{th_preview()} previews the content of a structure stored in the Threadle CLI
environment under the variable name \code{structure}.
}
\examples{
\dontshow{if (th_is_available()) withAutoprint(\{ # examplesIf}
th_start_threadle()

ns <- th_create_nodeset("ns", createnodes = 3)
th_preview(ns, maxlines = 10)
th_stop_threadle()
\dontshow{\}) # examplesIf}
}
