% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_remove_attr}
\alias{th_remove_attr}
\title{Remove a node attribute value}
\usage{
th_remove_attr(structure, nodeid, attrname)
}
\arguments{
\item{structure}{A \code{threadle_nodeset} or \code{threadle_network} object, or a character
string naming a structure in the Threadle CLI environment.}

\item{nodeid}{Node ID.}

\item{attrname}{Attribute name.}
}
\value{
\code{NULL}, invisibly.
}
\description{
\code{th_remove_attr()} removes an attribute value from a specific node
in a nodeset (or in the nodeset of the provided network).
}
