% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_remove_hyper}
\alias{th_remove_hyper}
\title{Remove a hyperedge from a two-mode layer}
\usage{
th_remove_hyper(network, layername, hypername)
}
\arguments{
\item{network}{A \code{threadle_network} object or a character string giving
the name of a network in the Threadle CLI environment.}

\item{layername}{Name of the two-mode layer.}

\item{hypername}{Name of the hyperedge to remove.}
}
\value{
\code{NULL}, invisibly.
}
\description{
\code{th_remove_hyper()} remove a hyperedge from a two-mode layer.
}
\examples{
\dontshow{if (th_is_available()) withAutoprint(\{ # examplesIf}
th_start_threadle()

ns <- th_create_nodeset("ns", createnodes = 3)
net <- th_create_network("net", ns)
th_add_layer(net, "t2", mode = 2)
th_add_hyper(net, "t2", "group1", nodes = c(1, 2))

th_remove_hyper(net, "t2", hypername = "group1")
th_get_all_hyperedges(net, "t2", offset = 0, limit = 10)
th_stop_threadle()
\dontshow{\}) # examplesIf}
}
