% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_shortest_path}
\alias{th_shortest_path}
\title{Calculate shortest path distance between two nodes
\code{th_shortest_path()} computes the shortest path distance from \code{node1id} to \code{node2id} in a network.}
\usage{
th_shortest_path(network, node1id, node2id, layername)
}
\arguments{
\item{network}{A \code{threadle_network} object or a character string giving
the name of a network in the Threadle CLI environment.}

\item{node1id}{Node ID of the first node.}

\item{node2id}{Node ID of the second node.}

\item{layername}{Optional layer name. If \code{NULL}, all layers are used.}
}
\value{
An integer scalar giving the shortest path distance.
}
\description{
Calculate shortest path distance between two nodes
\code{th_shortest_path()} computes the shortest path distance from \code{node1id} to \code{node2id} in a network.
}
\details{
By default, all layers are used. If \code{layername} is provided, the shortest path is
computed using that layer only.

Shortest path distance is directional: in directed layers, the distance from \code{node1id}
to \code{node2id} may differ from the distance in the reverse direction. For symmetric
layers, this distinction is moot.
}
\examples{
\dontshow{if (th_is_available()) withAutoprint(\{ # examplesIf}
th_start_threadle()

ns <- th_create_nodeset("ns", createnodes = 3)
net <- th_create_network("net", ns)
th_add_layer(net, "l1", mode = 1, directed = FALSE, valuetype = "binary", selfties = FALSE)
th_add_edge(net, "l1", node1id = 1, node2id = 2, value = 1)
th_add_edge(net, "l1", node1id = 2, node2id = 3, value = 1)

th_shortest_path(net, node1id = 1, node2id = 3, layername = "l1")
th_stop_threadle()
\dontshow{\}) # examplesIf}
}
