% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_stop_threadle}
\alias{th_stop_threadle}
\title{Stop the running Threadle CLI process}
\usage{
th_stop_threadle()
}
\value{
None; prints status messages.
}
\description{
Terminates the Threadle process previously started with `.start_threadle()`.
}
