% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_subnet}
\alias{th_subnet}
\title{Create a subnetwork from a network and nodeset}
\usage{
th_subnet(name, network, nodeset)
}
\arguments{
\item{name}{Name of the new network variable to create.}

\item{network}{A \code{threadle_network} object or a character string giving
the name of a network in the Threadle CLI environment.}

\item{nodeset}{A \code{threadle_nodeset} object or a character string giving
the name of a network in the Threadle CLI environment.}
}
\value{
A \code{threadle_network} object.
}
\description{
\code{th_subnet()} creates a new network based on an existing network, restricted to the
nodes in a specified nodeset.
}
\details{
This is useful for creating a network subset after creating a subset nodeset (e.g.,
with \code{th_filter()}).
}
\examples{
\dontshow{if (th_is_available()) withAutoprint(\{ # examplesIf}
th_start_threadle()

ns <- th_create_nodeset("ns", createnodes = 5)
net <- th_create_network("net", ns)

# make a subset nodeset containing nodes 1 and 2
ns_sub <- th_create_nodeset("ns_sub", createnodes = 0)
th_add_node(ns_sub, nodeid = 1)
th_add_node(ns_sub, nodeid = 2)

# create subnetwork induced by ns_sub
net_sub <- th_subnet("net_sub", net, ns_sub)
th_get_nbr_nodes(net_sub)
th_stop_threadle()
\dontshow{\}) # examplesIf}
}
