% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_symmetrize}
\alias{th_symmetrize}
\title{Create a symmetric version of a one-mode layer}
\usage{
th_symmetrize(
  network,
  layername,
  method = c("max", "min", "minnonzero", "average", "sum", "product"),
  newlayername = NULL
)
}
\arguments{
\item{network}{A \code{threadle_network} object or a character string giving the name of
a network in the Threadle CLI environment.}

\item{layername}{Name of the one-mode layer to symmetrize.}

\item{method}{Symmetrization method: \code{"max"}, \code{"min"}, \code{"minnonzero"}, \code{"average"}, \code{"sum"},
or \code{"product"}. Defaults to \code{"max"}.}

\item{newlayername}{Optional name for the newly created symmetrized layer.}
}
\value{
\code{NULL}, invisibly.
}
\description{
\code{th_symmetrize()} generates a symmetric version of a one-mode layer in a network.
}
\examples{
\dontshow{if (th_is_available()) withAutoprint(\{ # examplesIf}
th_start_threadle()

ns <- th_create_nodeset("ns", createnodes = 3)
net <- th_create_network("net", ns)
th_add_layer(net, "l1", mode = 1, directed = TRUE, valuetype = "valued", selfties = FALSE)
th_add_edge(net, "l1", node1id = 1, node2id = 2, value = 2)

th_symmetrize(net, "l1", method = "max", newlayername = "l1_sym")
th_get_edge(net, "l1_sym", node1id = 1, node2id = 2)
th_stop_threadle()
\dontshow{\}) # examplesIf}
}
