% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_unpack}
\alias{th_unpack}
\title{Convert a static layer back to a dynamic (editable) representation}
\usage{
th_unpack(network, layername = NULL)
}
\arguments{
\item{network}{A \code{threadle_network} object or a character string giving
the name of a network in the Threadle CLI environment.}

\item{layername}{Optional name of a single layer to unpack. If \code{NULL}, all
layers in the network are unpacked.}
}
\value{
\code{NULL}, invisibly.
}
\description{
\code{th_unpack()} converts one or all static layers in a network back to dynamic
layers, restoring full edit capability.
}
\examples{
\dontshow{if (th_is_available()) withAutoprint(\{ # examplesIf}
th_start_threadle()

ns <- th_create_nodeset("ns", createnodes = 4)
net <- th_create_network("net", ns)
th_add_layer(net, "l1", mode = 1, directed = FALSE, valuetype = "binary")
th_add_edge(net, "l1", node1id = 1, node2id = 2)
th_pack(net, layername = "l1")
th_unpack(net, layername = "l1")
th_stop_threadle()
\dontshow{\}) # examplesIf}
}
