% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_view}
\alias{th_view}
\title{View a structure by variable in the Threadle CLI environment}
\usage{
th_view(structure)
}
\arguments{
\item{name}{Name of the structure to view.}
}
\value{
CLI output.
}
\description{
View a structure by variable in the Threadle CLI environment
}
