% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{view}
\alias{view}
\title{View an object inside the Threadle CLI environment}
\usage{
view(name)
}
\arguments{
\item{name}{Name of the object to view.}
}
\value{
CLI output.
}
\description{
View an object inside the Threadle CLI environment
}
