% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{.send_command}
\alias{.send_command}
\title{Send a command to the Threadle CLI}
\usage{
.send_command(cmd)
}
\arguments{
\item{cmd}{A JSON command string to send to the Threadle CLI process.}
}
\value{
A parsed JSON response as a list.
}
\description{
Internal helper function used by all wrappers. Sends a JSON command and
waits for a complete one-line JSON response. It also support CLI command now.
}
\keyword{internal}
