% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{.th_call}
\alias{.th_call}
\title{Call the Threadle function}
\usage{
.th_call(
  cmd,
  args = NULL,
  assign = NULL,
  return = getOption("threadle.return", "payload"),
  .print_message = getOption("threadle.print_message", TRUE)
)
}
\description{
Call the Threadle function
}
