% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_add_aff}
\alias{th_add_aff}
\title{Add an affiliation to a two-mode layer}
\usage{
th_add_aff(
  network,
  layername,
  nodeid,
  hypername,
  addmissingnode = TRUE,
  addmissingaffiliation = TRUE
)
}
\arguments{
\item{network}{A \code{threadle_network} object or a character string giving
the name of a network in the Threadle CLI environment.}

\item{layername}{Layer name.}

\item{nodeid}{Node ID.}

\item{hypername}{Name of the hyperedge.}

\item{addmissingnode}{Logical; if \code{TRUE}, missing nodes are created automatically.}

\item{addmissingaffiliation}{Logical; if \code{TRUE}, missing affiliations are created automatically.}
}
\value{
\code{NULL}, invisibly.
}
\description{
\code{th_add_aff()} adds an affiliation (hyperedge membership) between a node and a hyperedge in a two-mode layer.
}
\examples{
\dontshow{if (th_is_available()) withAutoprint(\{ # examplesIf}
th_start_threadle()
ns <- th_create_nodeset("ns", createnodes = 5)
net <- th_create_network("net", ns)
th_add_layer(net, "t2", mode = 2)
th_add_hyper(net, "t2", "group1", c(1,2))
th_add_aff(net, "t2", nodeid = 1, hypername = "group1")
th_get_node_hyperedges(net, "t2", nodeid = 1)
th_stop_threadle()
\dontshow{\}) # examplesIf}
}
