% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_add_aff}
\alias{th_add_aff}
\title{Add an affiliation edge (from node to hyperedge) in a network layer}
\usage{
th_add_aff(
  network,
  layername,
  nodeid,
  hypername,
  addmissingnode = TRUE,
  addmissingaffiliation = TRUE
)
}
\description{
Add an affiliation edge (from node to hyperedge) in a network layer
}
