% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_add_edge}
\alias{th_add_edge}
\title{Add an edge to a network}
\usage{
th_add_edge(
  network,
  layername,
  node1id,
  node2id,
  value = 1,
  addmissingnodes = TRUE
)
}
\arguments{
\item{network}{A `threadle_network` object or a character string giving
the name of a network in the Threadle CLI environment.}

\item{layername}{Layer name.}

\item{node1id}{Node ID of the first vertex in the dyad.}

\item{node2id}{Node ID of the second vertex in the dyad.}

\item{value}{Edge value, defaults to 1.}

\item{addmissingnodes}{Logical; if `TRUE`, missing nodes are created and added. Defaults to `TRUE`.}
}
\value{
CLI output.
}
\description{
`th_add_edge()` adds an edge between two nodes in a layer.
}
