% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_add_edge}
\alias{th_add_edge}
\title{Add an edge to a network}
\usage{
th_add_edge(
  network,
  layername,
  node1id,
  node2id,
  value = 1,
  addmissingnodes = TRUE
)
}
\arguments{
\item{network}{A `threadle_nodeset` or `threadle_network` object.}

\item{id}{Node ID.}
}
\value{
CLI output.
}
\description{
Add an edge to a network
}
