% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_add_hyper}
\alias{th_add_hyper}
\title{Add a hyperedge to a two-mode layer}
\usage{
th_add_hyper(
  network,
  layername,
  hypername,
  nodes = c(),
  addmissingnodes = TRUE
)
}
\arguments{
\item{network}{A \code{threadle_network} object or a character string giving
the name of a network in the Threadle CLI environment.}

\item{layername}{Layer name.}

\item{hypername}{Name of the hyperedge.}

\item{nodes}{Vector of node to attach to the hyperedge. If \code{NULL},
creates the hyperedge without adding nodes.}

\item{addmissingnodes}{Logical; if \code{TRUE}, missing nodes are created and added. Defaults to \code{TRUE}.}
}
\value{
\code{NULL}, invisibly.
}
\description{
\code{th_add_hyper()} adds a hyperedge (affiliation) to the hyperedge set of a 2-mode layer.
}
\details{
If a hyperedge with the same name already exists, it is replaced. Duplicate node IDs in \code{nodes} are ignored.
}
