% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_add_hyper}
\alias{th_add_hyper}
\title{Add a hyperedge to a specified layer of a network (the hyperedge set of a 2-mode layer).}
\usage{
th_add_hyper(
  network,
  layername,
  hypername,
  nodes = c(1, 2, 3),
  addmissingnodes = FALSE
)
}
\arguments{
\item{network}{A `threadle_network` object.}

\item{id}{Node ID.}
}
\value{
CLI output.
}
\description{
Add a hyperedge to a specified layer of a network (the hyperedge set of a 2-mode layer).
}
