% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_add_layer}
\alias{th_add_layer}
\title{Adds/defines a relational layer in a network}
\usage{
th_add_layer(
  network,
  layername,
  mode,
  directed = FALSE,
  valuetype = c("binary", "valued"),
  selfties = FALSE
)
}
\arguments{
\item{network}{A `threadle_network` object or a character string giving
the name of a network in the Threadle CLI environment.}

\item{layername}{Layer name.}

\item{mode}{Layer mode (1 or 2).}

\item{directed}{Logical; whether ties are directed (only for 1-mode layers).}

\item{valuetype}{"binary" or "valued" (only for 1-mode layers).}

\item{selfties}{Logical; whether self-ties are allowed (only for 1-mode layers).}
}
\value{
CLI output.
}
\description{
Adds/defines a relational layer in a network
}
