% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_add_node}
\alias{th_add_node}
\title{Add a node to a nodeset}
\usage{
th_add_node(structure, nodeId)
}
\arguments{
\item{structure}{A \code{threadle_nodeset} or \code{threadle_network} object, or a character
string naming a structure in the Threadle CLI environment.}

\item{nodeId}{Node ID (unsigned integer).}
}
\value{
\code{NULL}, invisibly.
}
\description{
\code{th_add_node()} adds a node to a nodeset (or the nodeset of the provided network).
}
