% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_add_node}
\alias{th_add_node}
\title{Add a node to a nodeset (or network)}
\usage{
th_add_node(structure, id)
}
\arguments{
\item{structure}{A `threadle_nodeset` or `threadle_network` object, or a character
string naming a structure in the Threadle CLI environment.}

\item{id}{Node ID.}
}
\value{
CLI output.
}
\description{
Add a node to a nodeset (or network)
}
