% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_check_edge}
\alias{th_check_edge}
\title{Check if an edge exists between two nodes in a layer}
\usage{
th_check_edge(network, layername, node1Id, node2Id)
}
\arguments{
\item{network}{A \code{threadle_network} object or a character string giving
the name of a network in the Threadle CLI environment.}

\item{layername}{Layer name.}

\item{node1Id}{Node ID of the first node.}

\item{node2Id}{Node ID of the second node.}
}
\value{
A logical scalar indicating whether the specified tie exists.
}
\description{
\code{th_check_edge()} checks if an edge exists between two nodes in a layer, which can be either one-mode or two-mode.
}
