% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_check_edge}
\alias{th_check_edge}
\title{Check whether an edge exists in a layer}
\usage{
th_check_edge(network, layername, node1id, node2id)
}
\arguments{
\item{network}{A `threadle_network` object or a character string giving
the name of a network in the Threadle CLI environment.}

\item{layername}{Layer name.}

\item{node1id}{Node ID of the first vertex in the dyad.}

\item{node2id}{Node ID of the second vertex in the dyad.}
}
\value{
CLI output.
}
\description{
`th_check_edge()` checks whether an edge between
two nodes exists in the specified layer.
}
