% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_components}
\alias{th_components}
\title{Calculate connected components for the specified layer and network}
\usage{
th_components(network, layername, attrname)
}
\arguments{
\item{network}{A \code{threadle_network} object or a character string giving
the name of a network in the Threadle CLI environment.}

\item{attrname}{Attribute name.}

\item{layname}{Layer name.}
}
\value{
A named list with:
\describe{
\item{NbrComponents}{Integer of length 1; number of connected components.}
\item{ComponentSizes}{Integer vector; sizes of the connected components.}
}
}
\description{
\code{th_components()} computes connected components for the specified layer and network.
}
