% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_components}
\alias{th_components}
\title{Calculates the number of components for the specified layer and network.}
\usage{
th_components(network, layname, attrname)
}
\arguments{
\item{network}{A `threadle_network` object or a character string giving
the name of a network in the Threadle CLI environment.}

\item{layname}{Layer name.}

\item{attrname}{Attribute name.}
}
\value{
CLI output.
}
\description{
Calculates the number of components for the specified layer and network, and storing the result as an integer node attribute representing which component it is part of.
}
