% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_create_nodeset}
\alias{th_create_nodeset}
\title{Create a new nodeset in Threadle and assign it to variable 'name' in the Threadle CLI environment}
\usage{
th_create_nodeset(name, createnodes = 0)
}
\arguments{
\item{name}{Name of the R variable to assign in the CLI environment.}

\item{createnodes}{Number of nodes}
}
\value{
A `threadle_nodeset` object.
}
\description{
Create a new nodeset in Threadle and assign it to variable 'name' in the Threadle CLI environment
}
