% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_define_attr}
\alias{th_define_attr}
\title{Define an attribute for a nodeset (or network)}
\usage{
th_define_attr(
  structure,
  attrname,
  attrtype = c("int", "char", "float", "bool")
)
}
\arguments{
\item{structure}{A `threadle_nodeset` or `threadle_network` object, or a character
string naming a structure in the Threadle CLI environment.}

\item{attrname}{Attribute name.}

\item{attrtype}{Attribute type ("int","float", "char" or "bool")}
}
\value{
CLI output.
}
\description{
Define an attribute for a nodeset (or network)
}
