% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_degree}
\alias{th_degree}
\title{Calculate the degree centrality for a layer}
\usage{
th_degree(
  network,
  layername,
  attrname = NULL,
  direction = c("in", "out", "both")
)
}
\arguments{
\item{network}{A `threadle_network` object or a character string giving
the name of a network in the Threadle CLI environment.}

\item{layername}{Layer name.}

\item{attrname}{attribute name.}

\item{direction}{Edge direction: `"in"`, `"out"`, or `"both"`. Defaults to `"in"`.}
}
\value{
CLI output.
}
\description{
`th_degree()` computes the degree centrality for the specified network and layer.
}
