% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_degree}
\alias{th_degree}
\title{Calculates the degree centrality for the specified layer and network}
\usage{
th_degree(network, layername, attrname = NULL, direction = "in")
}
\description{
Calculates the degree centrality for the specified layer and network
}
