% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_delete}
\alias{th_delete}
\title{Delete a structure}
\usage{
th_delete(structure)
}
\arguments{
\item{structure}{A `threadle_nodeset` or `threadle_network` object, or a character
string naming a structure in the Threadle CLI environment.}
}
\value{
CLI output.
}
\description{
Delete a structure
}
