% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_delete_all}
\alias{th_delete_all}
\title{Delete all structures}
\usage{
th_delete_all()
}
\value{
\code{NULL}, invisibly.
}
\description{
\code{th_delete_all()} removes all current structures.
}
