% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_density}
\alias{th_density}
\title{Calculate the density of a layer}
\usage{
th_density(network, layername, samplesize = 200L)
}
\arguments{
\item{network}{A \code{threadle_network} object or a character string giving
the name of a network in the Threadle CLI environment.}

\item{layername}{Name of the layer for which density is computed.}

\item{samplesize}{Sample size for density estimation. If \code{NULL} (default),
the default is 200.}
}
\value{
A numeric scalar giving the layer density.
}
\description{
\code{th_density()} computes the density of a layer, treating all existing ties as binary.
}
\examples{
\dontshow{if (th_is_available()) withAutoprint(\{ # examplesIf}
th_start_threadle()

ns <- th_create_nodeset("ns", createnodes = 4)
net <- th_create_network("net", ns)
th_add_layer(net, "l1", mode = 1)
th_add_edge(net, "l1", node1id = 1, node2id = 2)
th_density(net, "l1")
th_stop_threadle()
\dontshow{\}) # examplesIf}
}
