% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_dichotomize}
\alias{th_dichotomize}
\title{dichotomize}
\usage{
th_dichotomize(
  network,
  layername,
  cond = "ge",
  threshold = 1,
  truevalue = 1,
  falsevalue = 0,
  newlayername = NULL
)
}
\description{
dichotomize
}
