% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_export}
\alias{th_export}
\title{Export a network to an external file format}
\usage{
th_export(network, format = "gexf", file, layername = NULL)
}
\arguments{
\item{network}{A \code{threadle_network} object or a character string giving
the name of a network in the Threadle CLI environment.}

\item{format}{Export format. Currently \code{"gexf"} is supported.}

\item{file}{Path to the output file.}

\item{layername}{Optional name of a specific layer to export. If \code{NULL},
all layers are included.}
}
\value{
\code{NULL}, invisibly.
}
\description{
\code{th_export()} exports a network to an external format such as GEXF
(readable by Gephi and other network tools).
}
\examples{
\dontshow{if (th_is_available()) withAutoprint(\{ # examplesIf}
th_start_threadle()

ns <- th_create_nodeset("ns", createnodes = 3)
net <- th_create_network("net", ns)
th_add_layer(net, "l1", mode = 1, directed = FALSE, valuetype = "binary")
th_add_edge(net, "l1", node1id = 1, node2id = 2)
tmp <- tempfile(fileext = ".gexf")
th_export(net, format = "gexf", file = tmp)
th_stop_threadle()
\dontshow{\}) # examplesIf}
}
