% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_filter}
\alias{th_filter}
\title{filter a nodeset to get a new filtered nodeset.}
\usage{
th_filter(name, nodeset, attrname, cond, attrvalue)
}
\description{
filter a nodeset to get a new filtered nodeset.
}
