% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_get_attr}
\alias{th_get_attr}
\title{Get the value of a node attribute for a nodeset (or network)}
\usage{
th_get_attr(structure, nodeid, attrname)
}
\arguments{
\item{structure}{A `threadle_nodeset` or `threadle_network` object, or a character
string naming a structure in the Threadle CLI environment.}

\item{nodeid}{Node ID.}

\item{attrname}{Attribute name.}
}
\value{
CLI output.
}
\description{
Get the value of a node attribute for a nodeset (or network)
}
