% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_get_edge}
\alias{th_get_edge}
\title{Get the value of an edge between two nodes}
\usage{
th_get_edge(network, layername, node1Id, node2Id)
}
\arguments{
\item{network}{A \code{threadle_network} object or a character string giving
the name of a network in the Threadle CLI environment.}

\item{layername}{Name of the layer to query (one-mode or two-mode).}

\item{node1Id}{Node ID of the first node.}

\item{node2Id}{Node ID of the second node.}
}
\value{
A numeric scalar giving the edge value (0 if no edge exists).
}
\description{
\code{th_get_edge()} gets the edge value between
two nodes in a specified layer.
}
