% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_get_hyperedge_nodes}
\alias{th_get_hyperedge_nodes}
\title{Get nodes affiliated with a hyperedge}
\usage{
th_get_hyperedge_nodes(network, layername, hypername)
}
\arguments{
\item{network}{A \code{threadle_network} object or a character string giving
the name of a network in the Threadle CLI environment.}

\item{layername}{Name of the two-mode layer to query.}

\item{hypername}{Name of the hyperedge.}
}
\value{
An integer vector of node IDs.
}
\description{
\code{th_get_hyperedge_nodes()} gets the node IDs affiliated with a hyperedge in a two-mode layer.
}
