% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_get_hyperedge_nodes}
\alias{th_get_hyperedge_nodes}
\title{Get nodes affiliated with a hyperedge}
\usage{
th_get_hyperedge_nodes(network, layername, hypername)
}
\arguments{
\item{network}{A \code{threadle_network} object or a character string giving
the name of a network in the Threadle CLI environment.}

\item{layername}{Name of the two-mode layer to query.}

\item{hypername}{Name of the hyperedge.}
}
\value{
An integer vector of node IDs.
}
\description{
\code{th_get_hyperedge_nodes()} gets the node IDs affiliated with a hyperedge in a two-mode layer.
}
\examples{
\dontshow{if (th_is_available()) withAutoprint(\{ # examplesIf}
th_start_threadle()

ns <- th_create_nodeset("ns", createnodes = 3)
net <- th_create_network("net", ns)
th_add_layer(net, "t2", mode = 2)
th_add_hyper(net, "t2", "group1", nodes = c(1, 2))
th_get_hyperedge_nodes(net, "t2", hypername = "group1")
th_stop_threadle()
\dontshow{\}) # examplesIf}
}
