% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_get_nbr_nodes}
\alias{th_get_nbr_nodes}
\title{Get the number of nodes in a structure}
\usage{
th_get_nbr_nodes(structure)
}
\arguments{
\item{structure}{A \code{threadle_nodeset} or \code{threadle_network} object, or a character
string naming a structure in the Threadle CLI environment.}
}
\value{
An integer scalar giving the number of nodes.
}
\description{
\code{th_get_nbr_nodes()} gets the number of nodes in the nodeset (or the nodeset of the provided network).
}
