% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_get_node_alters}
\alias{th_get_node_alters}
\title{Get alters of a node within a layer or across layers}
\usage{
th_get_node_alters(
  network,
  nodeId,
  layernames = "",
  direction = "both",
  unique = FALSE
)
}
\arguments{
\item{network}{A \code{threadle_network} object or a character string giving
the name of a network in the Threadle CLI environment.}

\item{nodeId}{Node ID.}

\item{layernames}{Optional layer names.
If \code{NULL}, alters are collected across all layers.}

\item{direction}{Which ties to count: \code{"both"} (default), \code{"in"}, or \code{"out"}.}

\item{unique}{Logical; if \code{TRUE}, deduplicate alter IDs. Defaults to \code{FALSE}.}
}
\value{
An integer vector giving alter node IDs.
}
\description{
Get alters of a node within a layer or across layers
}
