% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_get_nodeid_by_index}
\alias{th_get_nodeid_by_index}
\title{Get a node ID by index}
\usage{
th_get_nodeid_by_index(structure, index)
}
\arguments{
\item{structure}{Name of the structure (can be network or nodeset).}

\item{index}{Numeric index.}
}
\value{
The node ID.
}
\description{
Get a node ID by index
}
