% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_get_random_alter}
\alias{th_get_random_alter}
\title{Get a random alter for a node}
\usage{
th_get_random_alter(
  network,
  nodeid,
  layername = "",
  direction = c("both", "in", "out"),
  balanced = FALSE
)
}
\arguments{
\item{network}{Network name.}

\item{nodeid}{Node ID.}

\item{layername}{Optional layer. If left blank, will pick from all layers}

\item{direction}{Direction ("both"(default), "in", "out").}

\item{balanced}{Whether selection should be balanced for multiple layers.
Must be `TRUE` or `FALSE`. Defaults to `FALSE`.}
}
\value{
A node ID (numeric).
}
\description{
`th_get_random_alter()` get a random alter for a node.
}
