% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_get_random_node}
\alias{th_get_random_node}
\title{Get a random node from a structure}
\usage{
th_get_random_node(structure)
}
\arguments{
\item{structure}{Structure name.}
}
\value{
A node ID (numeric).
}
\description{
`th_get_random_node()` gets a random node.
}
