% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_import_layer}
\alias{th_import_layer}
\title{Import a layer into a network}
\usage{
th_import_layer(
  network,
  layername,
  file,
  format = c("edgelist", "matrix"),
  sep = "\\t",
  addmissingnodes = FALSE
)
}
\arguments{
\item{network}{Name of the network}

\item{layername}{Name of the layer to create inside `network`.}

\item{file}{Path to the input file.}

\item{format}{Input file format. One of `"edgelist"` or `"matrix"`.}

\item{sep}{Field separator used when `format = "edgelist"` (and for delimited
matrix formats, if applicable). Defaults to tab.}

\item{addmissingnodes}{Logical; if `TRUE`, create nodes referenced in the file
that are not yet present in `network`.}
}
\value{
Invisibly returns the CLI response.
}
\description{
`th_import_layer()` imports a layer from a file into an existing network.
}
