% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_info}
\alias{th_info}
\title{Retrieve meta information from a Threadle object}
\usage{
th_info(structure)
}
\arguments{
\item{structure}{A `threadle_nodeset` or `threadle_network` object, or a character
string naming a structure in the Threadle CLI environment.}
}
\value{
Parsed JSON or raw CLI text.
}
\description{
Retrieve meta information from a Threadle object
}
