% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_info}
\alias{th_info}
\title{Retrieve meta information from a Threadle object}
\usage{
th_info(structure, format = "json")
}
\arguments{
\item{structure}{A `threadle_nodeset` or `threadle_network` object.}

\item{format}{Output format ("json"(default) or "console").}
}
\value{
Parsed JSON or raw CLI text.
}
\description{
Retrieve meta information from a Threadle object
}
