% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_load_file}
\alias{th_load_file}
\title{Load a file into Threadle and assigns the structure(s) to the provided variable name in the Threadle CLI environment}
\usage{
th_load_file(name, file, type)
}
\arguments{
\item{name}{Name of the assigned variable in the Threadle CLI environment.}

\item{file}{File path.}

\item{type}{Type of structure ("network" or "nodeset").}
}
\value{
An object with class corresponding to the loaded type.
}
\description{
Load a file into Threadle and assigns the structure(s) to the provided variable name in the Threadle CLI environment
}
