% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_load_script}
\alias{th_load_script}
\title{Load and execute a Threadle CLI script}
\usage{
th_load_script(file)
}
\arguments{
\item{file}{Path to the script file.}
}
\value{
CLI output.
}
\description{
`th_load_script()` loads and executes a script file containing text CLI commands.
The backend will abort if it encounters an error in the script.
}
