% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_remove_aff}
\alias{th_remove_aff}
\title{Remove an affiliation from a two-mode layer}
\usage{
th_remove_aff(network, layername, nodeId, hypername)
}
\arguments{
\item{network}{A \code{threadle_network} object or a character string giving
the name of a network in the Threadle CLI environment.}

\item{layername}{Name of the two-mode layer.}

\item{nodeId}{Node ID.}

\item{hypername}{Name of the hyperedge.}
}
\value{
\code{NULL}, invisibly.
}
\description{
\code{th_remove_aff()} removes an affiliation between
a node and a hyperedge in a two-mode layer.
}
