% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_remove_aff}
\alias{th_remove_aff}
\title{Remove an affiliation (node -> hyperedge) from a layer}
\usage{
th_remove_aff(network, layername, nodeid, hypername)
}
\arguments{
\item{network}{A `threadle_network` object or a character string giving
the name of a network in the Threadle CLI environment.}

\item{layername}{Layer name.}

\item{nodeid}{Node ID.}

\item{hypername}{Name of the hyperedge.}
}
\value{
CLI output.
}
\description{
`th_remove_aff()` removes an affiliation between
a node and a hyperedge in the specified layer.
}
