% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_remove_aff}
\alias{th_remove_aff}
\title{Remove a structure}
\usage{
th_remove_aff(network, layername, nodeid, hypername)
}
\arguments{
\item{network}{A `threadle_network` object or a character string giving 
the name of a network in the Threadle CLI environment.}

\item{layername}{Name of the layer.}

\item{nodeid}{ID of the node.}

\item{hypername}{Name of the hyperedge.}

\item{structure}{A `threadle_nodeset` or `threadle_network`, or a character
string naming a structure in the Threadle CLI environment.}
}
\value{
CLI output.

CLI output.

CLI output.
}
\description{
Remove a structure
}
