% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_remove_attr}
\alias{th_remove_attr}
\title{Remove an attribute value from a node}
\usage{
th_remove_attr(structure, nodeid, attrname)
}
\arguments{
\item{structure}{A `threadle_nodeset` or `threadle_network` object, or a character
string naming a structure in the Threadle CLI environment.}

\item{nodeid}{Node ID.}

\item{attrname}{Attribute name.}
}
\value{
CLI output.
}
\description{
`th_remove_attr()` removes an attribute value for
a given node.
}
