% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_remove_edge}
\alias{th_remove_edge}
\title{Remove an edge from a layer}
\usage{
th_remove_edge(network, layername, node1id, node2id)
}
\arguments{
\item{network}{A `threadle_network` object or a character string giving 
the name of a network in the Threadle CLI environment.}

\item{layername}{The name of the layer.}

\item{node1id}{The ID of the first node.}

\item{node2id}{The ID of the second node.}
}
\value{
Invisibly returns the result of the remove operation.
}
\description{
`th_remove_edge` removes an edge between two
nodes in the specified layer.
}
