% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_remove_edge}
\alias{th_remove_edge}
\title{Remove an edge from a network layer}
\usage{
th_remove_edge(network, layername, node1id, node2id)
}
\arguments{
\item{network}{A network object or the name of a network variable.}

\item{layername}{The name of the layer.}

\item{node1id}{The ID of the first node.}

\item{node2id}{The ID of the second node.}
}
\value{
Invisibly returns the result of the remove operation.
}
\description{
Removes an edge between two nodes from a specified layer in a network.
}
