% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_remove_hyper}
\alias{th_remove_hyper}
\title{Remove a hyperedge from a 2-mode layer}
\usage{
th_remove_hyper(network, layername, hypername)
}
\arguments{
\item{network}{A `threadle_network` object or a character string giving
the name of a network in the Threadle CLI environment.}

\item{layername}{Layer name.}

\item{hypername}{Name of the hyperedge.}
}
\value{
CLI output.
}
\description{
`th_remove_hyper()` remove a hyperedge from the
specified layer.
}
