% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_remove_layer}
\alias{th_remove_layer}
\title{Remove a layer from a network}
\usage{
th_remove_layer(network, layername)
}
\arguments{
\item{network}{A \code{threadle_network} object or a character string giving
the name of a network in the Threadle CLI environment.}

\item{layername}{Name of the layer to remove.}
}
\value{
\code{NULL}, invisibly.
}
\description{
\code{th_remove_layer()} removes a layer from a network, including all ties in that layer.
}
