% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_remove_layer}
\alias{th_remove_layer}
\title{Remove a layer from a network}
\usage{
th_remove_layer(network, layername)
}
\arguments{
\item{network}{A `threadle_network` object or a character string giving
the name of a network in the Threadle CLI environment.}

\item{layername}{Name of the layer to be removed.}
}
\value{
CLI output.
}
\description{
`th_remove_layer()` removes a layer
(and its edges) from the network.
}
