% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_set_attr}
\alias{th_set_attr}
\title{Set a node attribute value}
\usage{
th_set_attr(structure, nodeid, attrname, attrvalue)
}
\arguments{
\item{structure}{A \code{threadle_nodeset} or \code{threadle_network} object, or a character
string naming a structure in the Threadle CLI environment.}

\item{nodeid}{Node ID.}

\item{attrname}{Attribute name.}

\item{attrvalue}{Attribute value to set.}
}
\value{
\code{NULL}, invisibly.
}
\description{
\code{th_set_attr()} sets the value of a node attribute for a specific node
in a nodeset (or the nodeset of the provided network).
}
