% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_set_attr}
\alias{th_set_attr}
\title{Set the value of a node attribute for a nodeset (or network)}
\usage{
th_set_attr(structure, nodeid, attrname, attrvalue)
}
\arguments{
\item{structure}{A `threadle_nodeset` or `threadle_network` object, or a character
string naming a structure in the Threadle CLI environment.}

\item{nodeid}{Node ID.}

\item{attrname}{Attribute name.}

\item{attrvalue}{Value to assign.}
}
\value{
CLI output.
}
\description{
Set the value of a node attribute for a nodeset (or network)
}
