% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_set_workdir}
\alias{th_set_workdir}
\title{Set working directory inside the Threadle CLI environment}
\usage{
th_set_workdir(dir)
}
\arguments{
\item{dir}{Path to the directory.}
}
\value{
CLI output as a character vector.
}
\description{
Set working directory inside the Threadle CLI environment
}
\examples{
\dontrun{
th_set_workdir("~/data")
}
}
