% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_setting}
\alias{th_setting}
\title{Set a Threadle backend setting}
\usage{
th_setting(name, value)
}
\arguments{
\item{name}{Setting name.}

\item{value}{Setting value (typically `"true"` or `"false"`).}
}
\value{
CLI output.
}
\description{
`th_setting()` changes setting in the Threadle backend.
}
