% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_start_threadle}
\alias{th_start_threadle}
\title{Start a Threadle CLI process}
\usage{
th_start_threadle(path = NULL)
}
\arguments{
\item{path}{Optional path to the Threadle CLI executable. If \code{NULL}, tries to
locate \code{threadle} on \code{PATH} via \code{\link[=Sys.which]{Sys.which()}}.}
}
\value{
The \code{processx} process object, invisibly.
}
\description{
Launches the Threadle CLI process executable and stores the process handle
in the global environment as \code{.threadle_proc}. The process is started in
silent mode (and optionally JSON mode).
}
\examples{
\dontshow{if (th_is_available()) withAutoprint(\{ # examplesIf}
th_start_threadle()

th_i()
th_stop_threadle()
\dontshow{\}) # examplesIf}
}
