% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_subnet}
\alias{th_subnet}
\title{Create a subnet from a network and a nodeset}
\usage{
th_subnet(name, network, nodeset)
}
\arguments{
\item{name}{Name of the new network variable to create.}

\item{network}{A `threadle_network` object or a character string giving
the name of a network in the Threadle CLI environment.}

\item{nodeset}{A `threadle_nodeset` object or a character string giving
the name of a network in the Threadle CLI environment.}
}
\value{
CLI output.
}
\description{
`th_subnet()` constructs a new network induced
by the nodes in `nodeset` and stores it under `name`.
}
