% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_symmetrize}
\alias{th_symmetrize}
\title{Create a symmetric version of a one-mode layer}
\usage{
th_symmetrize(
  network,
  layername,
  method = c("max", "min", "minnonzero", "average", "sum", "product"),
  newlayername = NULL
)
}
\arguments{
\item{network}{A \code{threadle_network} object or a character string giving the name of
a network in the Threadle CLI environment.}

\item{layername}{Name of the one-mode layer to symmetrize.}

\item{method}{Symmetrization method: \code{"max"}, \code{"min"}, \code{"minnonzero"}, \code{"average"}, \code{"sum"},
or \code{"product"}. Defaults to \code{"max"}.}

\item{newlayername}{Optional name for the newly created symmetrized layer.}
}
\value{
\code{NULL}, invisibly.
}
\description{
\code{th_symmetrize()} generates a symmetric version of a one-mode layer in a network.
}
