% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_sync_wd}
\alias{th_sync_wd}
\title{Synchronize Threadle working directory with the current R working directory}
\usage{
th_sync_wd()
}
\value{
\code{NULL}, invisibly.
}
\description{
Sets the working directory used by the Threadle process to the current R working directory returned by \code{\link[=getwd]{getwd()}}.
}
\details{
This is useful when invoking Threadle from threadleR, as the Threadle process
may have a different working directory than the R session.
}
\examples{
\dontshow{if (th_is_available()) withAutoprint(\{ # examplesIf}
th_start_threadle()

th_sync_wd()
th_stop_threadle()
\dontshow{\}) # examplesIf}
}
