% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_undefine_attr}
\alias{th_undefine_attr}
\title{Undefine a node attribute}
\usage{
th_undefine_attr(structure, attrname)
}
\arguments{
\item{structure}{A \code{threadle_nodeset} or \code{threadle_network} object, or a character
string naming a structure in the Threadle CLI environment.}

\item{attrname}{Attribute name to undefine.}
}
\value{
\code{NULL}, invisibly.
}
\description{
\code{th_undefine_attr()} removes the definition of a node attribute from a nodeset
(or the nodeset of the provided network).
}
\details{
Undefining an attribute also removes its values from all nodes that currently have
a value set for that attribute.
}
\examples{
\dontshow{if (th_is_available()) withAutoprint(\{ # examplesIf}
th_start_threadle()

ns <- th_create_nodeset("ns", createnodes = 3)
th_define_attr(ns, "grp", "int")
th_set_attr(ns, nodeid = 1, attrname = "grp", attrvalue = 1)
th_undefine_attr(ns, "grp")
th_info(ns)
th_stop_threadle()
\dontshow{\}) # examplesIf}
}
