% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threadle_client.R
\name{th_undefine_attr}
\alias{th_undefine_attr}
\title{Undefine an attribute on a structure}
\usage{
th_undefine_attr(structure, attrname)
}
\arguments{
\item{structure}{Name of the structure.}

\item{attrname}{Attribute name.}
}
\value{
CLI output.
}
\description{
`th_undefine_attr()` removes an attribute
from a nodeset or network.
}
